double everywhere
git-svn-id: svn+ssh://143.225.152.8/home/pasquale/SVN/signal-analyzer/branches/v2-development@31 bcba4c40-192e-4161-9d92-76493bb20e91

+++ src/main/java/com/dmp/signalanalyzer/filters/HighPassFilter.java
@@ -22,7 +22,7 @@ public class HighPassFilter extends SignalFilter{
       
       Signal previousPulse = firstItem;
       for (Signal pulse : inputSignal){
-          float filteredValue = (alpha * previousFiltered.getValue()) +
+          double filteredValue = (alpha * previousFiltered.getValue()) +
                   (beta * (pulse.getValue() - previousPulse.getValue()));
           
           previousPulse = pulse;
