+++ src/main/java/com/dmp/signalanalyzer/filters/NinetiethPercentSelector.java
@@ -17,10 +17,10 @@ public class NinetiethPercentSelector extends SignalFilter {
       Signal filteredSignal = new Signal();
       // 1 compute 90°perc 
       percentile.setQuantile(quantile);
-      float percentileVal = (float) percentile.evaluate(signal.toDoubleValuesArray());
+      double percentileVal = percentile.evaluate(signal.toDoubleValuesArray());
       
       // 2 select pulses >= 90° Perc
-      float filteredValue = 0; //{0,1}
+      double filteredValue = 0; //{0,1}
       for (Signal pulse : signal){
          filteredValue = pulse.getValue() >= percentileVal ? pulse.getValue() : 0;
          filteredSignal.addComponent(new Signal(pulse.getTStart(),pulse.getTStop(), filteredValue));
