+++ src/main/java/com/dmp/signalanalyzer/filters/UnbiasFilter.java
@@ -10,7 +10,7 @@ public class UnbiasFilter extends SignalFilter {
 
     public Signal filter(Signal inputSignal) {
         Signal analysis = new Signal();
-        float prevAvg = 0;
+        double prevAvg = 0;
         int index = 0;
         for (Signal pulse : inputSignal) {
             prevAvg = moovingAverage(prevAvg, index, pulse.getValue());
@@ -22,9 +22,9 @@ public class UnbiasFilter extends SignalFilter {
         return analysis;
     }
 
-    private float moovingAverage(float prevAvg, int index, float value) {
+    private double moovingAverage(double prevAvg, int index, double value) {
         // E4*VALUE(A5-1)/A5+D5*VALUE(1/A5)
-        float movAvg = value;
+        double movAvg = value;
 
         if (index > 0) {
             movAvg =
