+++ src/main/java/com/dmp/signalanalyzer/filters/windowed/WindowedMedianAnalysis.java
@@ -13,9 +13,9 @@ public class WindowedMedianAnalysis extends WindowedSignalFilter {
    private static Percentile percentile = new Percentile();
 
    
-   public float getSingleWindowValue(Signal window) {
+   public double getSingleWindowValue(Signal window) {
       percentile.setQuantile(quantile);
-      float percentileVal = (float) percentile.evaluate(window.toDoubleValuesArray());
+      double percentileVal = percentile.evaluate(window.toDoubleValuesArray());
       return percentileVal;
    }
