+++ src/main/java/com/dmp/signalanalyzer/filters/windowed/WindowedNinetiethPercentileAnalysis.java
@@ -12,9 +12,9 @@ public class WindowedNinetiethPercentileAnalysis extends WindowedSignalFilter {
    private static final double quantile = 90.;
    private static Percentile percentile = new Percentile();
 
-   public float getSingleWindowValue(Signal window) {
+   public double getSingleWindowValue(Signal window) {
       percentile.setQuantile(quantile);
-      float percentileVal = (float) percentile.evaluate(window.toDoubleValuesArray());
+      double percentileVal = percentile.evaluate(window.toDoubleValuesArray());
       return percentileVal;
    }
 }
